Refactoring -- add TermuxEnvironment export termux Environment
